+++ src/Integration.UnitTests/MefServices/ActiveSolutionBoundTrackerTests.cs
@@ -25,6 +25,7 @@ using Microsoft.VisualStudio.Shell.Interop;
 using Moq;
 using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Core.Binding;
+using SonarLint.VisualStudio.Core.ConfigurationScope;
 using SonarLint.VisualStudio.Infrastructure.VS.Roslyn;
 using SonarLint.VisualStudio.TestInfrastructure;
 using SonarQube.Client;
